+++ src/features/common/components/index.ts
@@ -1,6 +1,3 @@
 export * from "./BackButton";
-export * from "./layout/Container";
-export * from "./Form";
 export * from "./Loader";
 export * from "./Modal";
-export * from "./layout/PageWrapper";
