+++ src/features/common/components/layout/index.ts
@@ -1,1 +1,2 @@
 export * from "./PageLayout";
+export * from "./Container";
