+++ src/features/food/components/Food/components/EditFood.tsx
@@ -6,8 +6,9 @@ import { api } from "~/utils/api";
 import { editFood } from "~/utils/schemas/food";
 import type { ActionProps } from "../types";
 import { ACTIONS } from "../constants";
-import { Modal, Input, Textarea } from "~/features/common/components";
+import { Modal } from "~/features/common/components";
 import { SelectFoodType } from "./SelectFoodType";
+import { Input, Textarea } from "~/features/common/components/Form";
 
 export const EditFood: React.FC<ActionProps> = ({
   data: food,
